fix(introduce new fsxa-api version): introduce new fsxa-api and refactor code to match new structure
a new version of the fsxa-api was released, that enables universal usage on server and browser. We
upgraded the fsxa-api package and refactored our code to match the new interface

BREAKING CHANGE: the fsxa-api is now working with the native fetch api, which is not available on
server side. if you are using the pattern-library make sure that you are polyfilling fetch through
cross-fetch for example.