feat(routing): added util method to replace FS formatting in richtexts with working links (#3)
* feat(routing): added util method to replace FS formatting in richtexts with working links

FirstSpirit is adding special formats for link templates into the richtext output that have to be
parsed and replaced to work as links in the application content. The `createLinksInRichText`-method
(provided by the FSXABaseComponent) will replace the special formats with working links. Make sure
that you transform the html-content received through the api with the newly provided method.