Feature/refactor (#4)
feat(datasets / devmode / refactor): did a major refactor, while adding full dataset-support

We've fully enabled the support for datasets inside of the pattern-library. This includes dynamic
routing and section-mapping. We've also refactored the pattern-library to slim down the bundle and
to remove unused or mislocated files. The pattern-library is now fully relying on components that
will be passed to it instead of providing a set of default components

- add full sfc compatibility
- add dynamic dataset support
- add new FSXAApp wrapper
- add refresh on navigation change
- remove default components
- add new and enhanced devMode
- add rudimentary examples

BREAKING CHANGE: - We removed the default components. Every component must now be mapped in your
project. (we've fully dropped fsxa-ui as dependency)
- You now have to use the FSXAApp as
root-wrapper.
- The interface for providing your component map slightly changed. You now can pass
all of your components through a single property.