feat(richtext): we upgraded the fsxa-api dependency and switched to the new RichText JSON structure
You can now import the `FSXARichText` component to render RichText-Nodes received through the
fsxa-api. This component will map the JSON to RichText-Components you provided through the
`components.richtext` map. Make sure to extend the `FSXABaseRichTextElement` to get access to the
passed properties. If no component was found and the devMode is active, you will see the
DevMode-Info you are used to. We also added the `pageId` prop back to the `FSXABaseLayout`.