fix(fragments): we've removed vue-fragment as dependency since this can break SSR
We used vue-fragment for returning multiple nodes as root. This works very well but breaks SSR
because the DOM is manipulated after rendering. Before SSR is finished a div will be rendered. This
can lead to invalid html errors, when used inside of p or span tags. We now removed the usage of
vue-fragment and switched to a functional component for rendering richtext in vue-sfc formats. Check
the devMode for more information.