fix(rendering): reduced wrapper divs as much as possible and added styling to required ones
We've removed some wrapper divs that could possibly make it harder to style the application. If it
was not possible to remove it, we added the w-full and h-full classes, so that it would not
interfere with your layout. We've also introduced the meta property to the FSXABaseSection. This
meta property will contain additional information about your section. For example a dataset section
will receive its route and type as meta data.