feat(tpp-snap): remove the bundled tpp-snap version and dynamically load specified version (#6)
You can now pass the version of the fs-tpp-api which should be used. This gives you more flexibility
to use new features, when the standard is not updated yet or helps you, when your server is running
an older version. If you do not specify a version the default will be 2.2.1 for now. If this
changes, we will inform you in our release notes.