feat(i18n): add availableLocales as property to the FSXAApp
With the `availableLocales` property you are able to define all locales that should be available in
your application. *Attention*: This will automatically load the navigation data of all locales. So
if you want to support a lot of languages with your application, it is better not to use this
option. If availableLocales is undefined, the application will dynamically load the required
navigation data.