feat: add NavigationFilter (#27)
* feat: update dependency for fsxa-api and adjust usages of fsxa-api classes (#17)
* fix: implement solution to deal with trailing slashes (#18)
* chore(TPP_SNAP): update DEFAULT_TPP_SNAP_VERSION to 2.4.0

The store was enhanced, so that the navigation and an auth object can be
changed. This can be used to show personalized content.

BREAKING CHANGE: The original fsxa-api class was removed and the new ones FSXAProxyApi and
FSXARemoteApi are used. Their have a slightly different, but better, method signatures. For
more information, please read the [migration guide in the CHANGELOG of the FSXA-API](https://github.com/e-Spirit/fsxa-api/blob/master/CHANGELOG.md#migration-guide).